+++ assets/case_event/copy_select2.js
@@ -1,0 +1,6 @@
+/* global $ */
+
+window.addEventListener('load', function () {
+  const $selector = $('#case_event_copy_cases')
+  $selector.select2()
+})