+++ assets/document/copy_select2.js
@@ -1,0 +1,6 @@
+/* global $ */
+
+window.addEventListener('load', function () {
+  const $selector = $('#copy_document_form_cases')
+  $selector.select2()
+})